refactor(file-storage): split store.rs and service.rs to ≤600 SLOC per file
`store.rs` (1175 SLOC) and `service.rs` (1014 SLOC) were the only files over the
600-SLOC budget. Split both so every source file is ≤600 (max is now dto.rs at
491); the struct/fields/constructor stay in `mod.rs` and public paths are
unchanged (`…::store::Store`, `…::service::FileService`).

- `store.rs` → `store/{mod,files,versions,metadata,policy,multipart,lifecycle,traits}.rs`
  (module-path split by the concern boundaries already marked in the file); each ≤280.
- `service.rs` → `service/{mod,create,finalize,read,update,delete,versioning,transfer,migration,data_plane_port}.rs`;
  `create.rs`/`migration.rs` are cohesive extractions, the rest a module-path split; each ≤248.

Trade-off (accepted, for the SLOC budget): splitting `impl FileService` across
submodules adds `use super::FileService` fan-in edges, so `service/mod.rs` HK
rises (1.54M → 2.19M) and shared `error`/`policy` fan-in grows; `store/mod.rs`
HK falls (794K → 335K) as its fan-out distributes across subfiles. Behaviour,
public API, gear wiring, and REST routes unchanged.

168 tests pass, clippy (-D warnings) + fmt clean.
