+++ gears/file-storage/file-storage/src/domain/service/delete.rs
@@ -1,0 +1,169 @@
+//! File and version deletion flows.
+
+use toolkit_security::{AccessScope, SecurityContext};
+use uuid::Uuid;
+
+use crate::domain::audit::AuditOperation;
+use crate::domain::authz::actions;
+use crate::domain::error::DomainError;
+use crate::domain::etag;
+use crate::domain::service::FileService;
+use crate::infra::external_clients::UsageDelta;
+
+impl FileService {
+    // ── delete ──────────────────────────────────────────────────────────────────
+
+    /// `DELETE /files/{id}`: remove the file and all versions (FK cascade) under
+    /// an `If-Match` content-ETag precondition, then best-effort delete the
+    /// backend blobs. `If-Match` is **required** (see api.md §DELETE); pass `"*"`
+    /// to delete unconditionally when the ETag is unknown.
+    ///
+    /// @cpt-cf-file-storage-fr-audit-trail
+    pub async fn delete_file(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+        if_match: Option<&str>,
+    ) -> Result<(), DomainError> {
+        let prefetch = Self::tenant_scope(ctx);
+        let file = self.store.require_file(&prefetch, file_id).await?;
+        let _scope = self
+            .authorizer
+            .authorize(ctx, actions::DELETE, &file.gts_file_type, Some(file_id))
+            .await?;
+
+        // Validate the If-Match precondition against the current content ETag.
+        let current_etag = etag::etag_for(&file);
+        match if_match {
+            None => {
+                return Err(DomainError::precondition_failed(
+                    "If-Match is required to delete a file",
+                ));
+            }
+            Some(m) => {
+                let m = m.trim();
+                if m != "*" && Some(m) != current_etag.as_deref() {
+                    return Err(DomainError::precondition_failed(
+                        "If-Match does not match the current content ETag",
+                    ));
+                }
+            }
+        }
+
+        self.delete_file_inner(ctx, file_id).await
+    }
+
+    /// Inner (unconditional) file deletion: authorization and If-Match must have
+    /// already been checked by the caller. Collects versions, removes the DB row
+    /// (and FK children via cascade), then best-effort-deletes all backend blobs.
+    ///
+    /// @cpt-cf-file-storage-fr-audit-trail
+    pub(super) async fn delete_file_inner(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+    ) -> Result<(), DomainError> {
+        // Authorization has already been verified by callers; use allow_all() for
+        // the DB scope — the tenant boundary was enforced by require_file() above.
+        let scope = AccessScope::allow_all();
+
+        // Collect backend blobs before the metadata row (and FK children) vanish.
+        let versions = self.store.list_versions(file_id).await?;
+
+        // @cpt-cf-file-storage-fr-audit-trail
+        let audit = Self::audit_ok(
+            ctx,
+            Some(file_id),
+            AuditOperation::DeleteFile,
+            serde_json::json!({ "version_count": versions.len() }),
+        );
+
+        // @cpt-cf-file-storage-fr-file-events
+        // We need the file's tenant/owner for the event payload; fetch before deletion.
+        let file_meta = self.store.get_file(&scope, file_id).await?;
+        let (event_tenant, event_owner) = file_meta.as_ref().map_or_else(
+            || (ctx.subject_tenant_id(), Uuid::nil()),
+            |f| (f.tenant_id, f.owner_id),
+        );
+        let event = Some(Self::make_file_event(
+            event_tenant,
+            event_owner,
+            file_id,
+            "file.deleted",
+            serde_json::json!({ "version_count": versions.len() }),
+        ));
+
+        let removed = self
+            .store
+            .delete_file_with_event(&scope, file_id, audit, event)
+            .await?;
+        if !removed {
+            return Err(DomainError::file_not_found(file_id));
+        }
+
+        // @cpt-cf-file-storage-fr-usage-reporting
+        let total_bytes: i64 = versions.iter().map(|v| v.size).sum();
+        self.report_usage(UsageDelta {
+            tenant_id: event_tenant,
+            owner_id: event_owner,
+            bytes_delta: -total_bytes,
+            file_count_delta: -1,
+        });
+
+        // Best-effort backend cleanup; a failure degrades to an orphan (P2 GC).
+        for v in versions {
+            self.best_effort_blob_delete(&v.backend_id, &v.backend_path)
+                .await;
+        }
+        Ok(())
+    }
+
+    /// Delete a single version (and its backend blob). Deleting the only version
+    /// is equivalent to deleting the file.
+    ///
+    /// @cpt-cf-file-storage-fr-audit-trail
+    pub async fn delete_version(
+        &self,
+        ctx: &SecurityContext,
+        file_id: Uuid,
+        version_id: Uuid,
+    ) -> Result<(), DomainError> {
+        let prefetch = Self::tenant_scope(ctx);
+        let file = self.store.require_file(&prefetch, file_id).await?;
+        let _scope = self
+            .authorizer
+            .authorize(ctx, actions::DELETE, &file.gts_file_type, Some(file_id))
+            .await?;
+
+        let all = self.store.list_versions(file_id).await?;
+        if all.len() <= 1 {
+            // Last version → delete the whole file. Authorization has already been
+            // checked above; skip the If-Match gate (delete_version has its own
+            // contract — no If-Match on DELETE /files/{id}/versions/{vid}).
+            return self.delete_file_inner(ctx, file_id).await;
+        }
+        let Some(version) = all.into_iter().find(|v| v.version_id == version_id) else {
+            return Err(DomainError::version_not_found(file_id, version_id));
+        };
+        if file.content_id == Some(version_id) {
+            return Err(DomainError::conflict(
+                "cannot delete the current version; bind another version first",
+            ));
+        }
+
+        // @cpt-cf-file-storage-fr-audit-trail
+        let audit = Self::audit_ok(
+            ctx,
+            Some(file_id),
+            AuditOperation::DeleteVersion,
+            serde_json::json!({ "version_id": version_id }),
+        );
+
+        self.store
+            .delete_version(file_id, version_id, audit)
+            .await?;
+        self.best_effort_blob_delete(&version.backend_id, &version.backend_path)
+            .await;
+        Ok(())
+    }
+}
